requirements.txt
transformers==4.42.3
timm==0.9.16
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
Pillow==10.3.0
tqdm==4.66.3 # THIS LINE HAS CHANGED
accelerate==0.31.0
kaggle==1.6.14
datasets==2.20.0